audioresample: changed num_gap_samples, num_nongap_samples from guint32 to guint64 so that gaps of greater than or equal to 2^32 samples do not cause integer overflow

+++ gst/audioresample/gstaudioresample.h
@@ -64,8 +64,8 @@ struct _GstAudioResample {
   guint64 samples_in;
   guint64 samples_out;
   
-  guint num_gap_samples;
-  guint num_nongap_samples;
+  guint64 num_gap_samples;
+  guint64 num_nongap_samples;
 
   gint channels;
   gint inrate;
